New upstream version 4.0.12

+++ src/mongo/util/safe_num-inl.h
@@ -1,28 +1,31 @@
-/*    Copyright 2013 10gen Inc.
+
+/**
+ *    Copyright (C) 2018-present MongoDB, Inc.
  *
- *    This program is free software: you can redistribute it and/or  modify
- *    it under the terms of the GNU Affero General Public License, version 3,
- *    as published by the Free Software Foundation.
+ *    This program is free software: you can redistribute it and/or modify
+ *    it under the terms of the Server Side Public License, version 1,
+ *    as published by MongoDB, Inc.
  *
  *    This program is distributed in the hope that it will be useful,
  *    but WITHOUT ANY WARRANTY; without even the implied warranty of
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *    GNU Affero General Public License for more details.
+ *    Server Side Public License for more details.
  *
- *    You should have received a copy of the GNU Affero General Public License
- *    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *    You should have received a copy of the Server Side Public License
+ *    along with this program. If not, see
+ *    <http://www.mongodb.com/licensing/server-side-public-license>.
  *
  *    As a special exception, the copyright holders give permission to link the
  *    code of portions of this program with the OpenSSL library under certain
  *    conditions as described in each individual source file and distribute
  *    linked combinations including the program with the OpenSSL library. You
- *    must comply with the GNU Affero General Public License in all respects
- *    for all of the code used other than as permitted herein. If you modify
- *    file(s) with this exception, you may extend this exception to your
- *    version of the file(s), but you are not obligated to do so. If you do not
- *    wish to do so, delete this exception statement from your version. If you
- *    delete this exception statement from all source files in the program,
- *    then also delete it in the license file.
+ *    must comply with the Server Side Public License in all respects for
+ *    all of the code used other than as permitted herein. If you modify file(s)
+ *    with this exception, you may extend this exception to your version of the
+ *    file(s), but you are not obligated to do so. If you do not wish to do so,
+ *    delete this exception statement from your version. If you delete this
+ *    exception statement from all source files in the program, then also delete
+ *    it in the license file.
  */
 
 #pragma once
@@ -41,11 +44,11 @@ inline SafeNum& SafeNum::operator=(const SafeNum& rhs) {
     return *this;
 }
 
-inline SafeNum::SafeNum(int num) : _type(NumberInt) {
+inline SafeNum::SafeNum(int32_t num) : _type(NumberInt) {
     _value.int32Val = num;
 }
 
-inline SafeNum::SafeNum(long long int num) : _type(NumberLong) {
+inline SafeNum::SafeNum(int64_t num) : _type(NumberLong) {
     _value.int64Val = num;
 }
 
@@ -70,7 +73,7 @@ inline SafeNum SafeNum::operator+(const SafeNum& rhs) const {
 }
 
 inline SafeNum& SafeNum::operator+=(const SafeNum& rhs) {
-    return * this = addInternal(*this, rhs);
+    return *this = addInternal(*this, rhs);
 }
 
 inline SafeNum SafeNum::operator*(const SafeNum& rhs) const {
@@ -78,7 +81,7 @@ inline SafeNum SafeNum::operator*(const SafeNum& rhs) const {
 }
 
 inline SafeNum& SafeNum::operator*=(const SafeNum& rhs) {
-    return * this = mulInternal(*this, rhs);
+    return *this = mulInternal(*this, rhs);
 }
 
 inline SafeNum SafeNum::bitAnd(const SafeNum& rhs) const {
@@ -90,7 +93,7 @@ inline SafeNum SafeNum::operator&(const SafeNum& rhs) const {
 }
 
 inline SafeNum& SafeNum::operator&=(const SafeNum& rhs) {
-    return * this = bitAnd(rhs);
+    return *this = bitAnd(rhs);
 }
 
 inline SafeNum SafeNum::bitOr(const SafeNum& rhs) const {
@@ -102,7 +105,7 @@ inline SafeNum SafeNum::operator|(const SafeNum& rhs) const {
 }
 
 inline SafeNum& SafeNum::operator|=(const SafeNum& rhs) {
-    return * this = bitOr(rhs);
+    return *this = bitOr(rhs);
 }
 
 inline SafeNum SafeNum::bitXor(const SafeNum& rhs) const {
@@ -114,7 +117,7 @@ inline SafeNum SafeNum::operator^(const SafeNum& rhs) const {
 }
 
 inline SafeNum& SafeNum::operator^=(const SafeNum& rhs) {
-    return * this = bitXor(rhs);
+    return *this = bitXor(rhs);
 }
 
 inline bool SafeNum::isValid() const {
